+++ js/utils/get-random-int-inclusive.js
@@ -1,0 +1,8 @@
+//Взял с MDN
+function getRandomIntInclusive(min, max) {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min + 1)) + min; //Максимум и минимум включаются
+}
+
+export {getRandomIntInclusive};
